Added EncodeService & IBitcoinNetwork.

+++ Enterprise/Bitcoin/CafeLib.Bitcoin.Shared/Encoding/HexReverseEncoder.cs
@@ -54,7 +54,7 @@ namespace CafeLib.Bitcoin.Shared.Encoding
                 var b = CharToNibble(hex[i++]);
 
                 if (a == -1 || b == -1) return false;
-                bytes[--j] = (byte)((a << 4) | (int)b);
+                bytes[--j] = (byte)((a << 4) | b);
             }
 
             return true;
